+++ crates/matrix-sdk/tests/integration/client.rs
@@ -771,7 +771,7 @@ async fn test_encrypt_room_event() {
     .cast();
 
     let push_action_ctx =
-        room.push_action_ctx().await.expect("We should be able to get the push action context");
+        room.push_context().await.expect("We should be able to get the push action context");
     let timeline_event = room
         .decrypt_event(&event, &push_action_ctx)
         .await